requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2021.5.30
lxml==4.6.3
Pillow==8.3.1
pkg_resources==0.0.0
python-docx==0.8.11
python-telegram-bot==13.7
pytz==2021.1
six==1.16.0
tornado==6.1
tzlocal==2.1
